Moved build/runtime.go to client/runtime
